Add files via upload
"To-Do Compressed.html" contains the full Javascript and CSS included in the html head. A new file can be created from scratch by typing into the text area and then hitting Enter (or clicking the button below) and then working with each Task that appears below. Whatever file is created can be downloaded by hitting the download button at the top-right of the page. To assign the name to the document that will be downloaded, simply type into the text input next to the button, otherwise a default name will be created. Additionally, "Todo to-dos.json" is a file that has already been created with the app and can be selected for upload by clicking on the 'Choose File' button under 'Upload Task File'.